+++ common/api/button.ts
@@ -1,3 +1,13 @@
+export interface TapOptions {
+    double?: boolean;
+    hold?: boolean;
+}
+
+export interface ClickOptions {
+    double?: boolean;
+    timeout?: number;
+}
 export interface ButtonLike {
-    click(): void | Promise<void>;
+    click(options?: ClickOptions): void | Promise<void>;
+    tap(options?: TapOptions): void | Promise<void>;
 }