+++ common/api/form.ts
@@ -1,0 +1,4 @@
+export interface Form<T> {
+    submit(): void | Promise<void>;
+    fill?(model: T): void | Promise<void>;
+}