+++ common/api/index.ts
@@ -1,4 +1,0 @@
-export * from './button';
-export * from './image';
-export * from './label';
-export * from './link';