+++ common/api/label.ts
@@ -1,1 +1,3 @@
-export interface LabelLike { }
+export interface LabelLike {
+    for?: string | Promise<string>;
+}