+++ common/api/script.ts
@@ -1,0 +1,4 @@
+export interface ScriptLike {
+    src?: string | Promise<string>;
+    getScriptFromUrl(url: string): Promise<string>;
+}