+++ common/label.ts
@@ -1,4 +1,8 @@
 import { Component } from "../src/component";
 import { LabelLike } from "./api/label";
 
-export class Label extends Component implements LabelLike { }
+export class Label extends Component implements LabelLike {
+    get for() {
+        return this.getAttribute('for');
+    }
+}