+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.1.8
+- add more common components(`Video`, `Script`, `Select`, `TextArea`, `Source`, `Input`)
+- add api for common components
+- add jsdoc from protractor
+- add input types
 # 1.1.0
 - Add common components(`Button`, `Image`, `Label`, `Link`)
 - add typings for common components